Align examples and JSON contract tests
Correct workflow, configuration, and JSON documentation contracts; make
example output directories explicit order-only dependencies; verify documented
artefacts; and split logging integration tests by feature.
