+++ docs/sample-netsuke.toml
@@ -18,7 +18,7 @@
 # Locale for CLI messages, for example `en-US` or `es-ES`.
 # locale = "en-US"
 
-# Emit machine-readable JSON results or diagnostics.
+# Emit one versioned JSON document: a result on success or diagnostic on failure.
 # json = false
 
 # Never read interactive input. This must remain true until an explicit
